+++ python/ray/autoscaler/command_runner.py
@@ -1,80 +1,4 @@
-from shlex import quote
 from typing import Any, List, Tuple, Dict, Optional
-import json
-import os
-import sys
-
-# How long to wait for a node to start, in seconds
-NODE_START_WAIT_S = 300
-HASH_MAX_LENGTH = 10
-KUBECTL_RSYNC = os.path.join(
-    os.path.dirname(os.path.abspath(__file__)), "kubernetes/kubectl-rsync.sh")
-
-_config = {"use_login_shells": True, "silent_rsync": True}
-
-
-def is_rsync_silent():
-    return _config["silent_rsync"]
-
-
-def set_rsync_silent(val):
-    """Choose whether to silence rsync output.
-
-    Most commands will want to list rsync'd files themselves rather than
-    print the default rsync spew.
-    """
-    _config["silent_rsync"] = val
-
-
-def is_using_login_shells():
-    return _config["use_login_shells"]
-
-
-def set_using_login_shells(val):
-    """Choose between login and non-interactive shells.
-
-    Non-interactive shells have the benefit of receiving less output from
-    subcommands (since progress bars and TTY control codes are not printed).
-    Sometimes this can be significant since e.g. `pip install` prints
-    hundreds of progress bar lines when downloading.
-
-    Login shells have the benefit of working very close to how a proper bash
-    session does, regarding how scripts execute and how the environment is
-    setup. This is also how all commands were ran in the past. The only reason
-    to use login shells over non-interactive shells is if you need some weird
-    and non-robust tool to work.
-
-    Args:
-        val (bool): If true, login shells will be used to run all commands.
-    """
-    _config["use_login_shells"] = val
-
-
-def _with_environment_variables(cmd: str,
-                                environment_variables: Dict[str, object]):
-    """Prepend environment variables to a shell command.
-
-    Args:
-        cmd (str): The base command.
-        environment_variables (Dict[str, object]): The set of environment
-            variables. If an environment variable value is a dict, it will
-            automatically be converted to a one line yaml string.
-    """
-
-    as_strings = []
-    for key, val in environment_variables.items():
-        val = json.dumps(val, separators=(",", ":"))
-        s = "export {}={};".format(key, quote(val))
-        as_strings.append(s)
-    all_vars = "".join(as_strings)
-    return all_vars + cmd
-
-
-def _with_interactive(cmd):
-    force_interactive = (
-        f"true && source ~/.bashrc && "
-        f"export OMP_NUM_THREADS=1 PYTHONWARNINGS=ignore && ({cmd})")
-    return ["bash", "--login", "-c", "-i", quote(force_interactive)]
 
 
 class CommandRunnerInterface:
